russia is republic in federation.
one republic is variable. federation is many republics, areay or maybe function from variables. but moscow in  what republic? Russia actually a republic in federation.real translation is federation of russia or federation of republic (of) russia. there many areas not designated as a free state, republic, but as if as part of federations.